fix: compilation error

+++ src/renderer/compositor.cpp
@@ -28,8 +28,10 @@ namespace System::Render
         typedef uint32_t assets_id;
         assets_id get_assets_id(const char *name);
         assets_id load_sprite(const char *path, const char *name, size_t width, size_t height);
-        assets_id load_vertex_shader(const char *path, const char *name, InputAttributeDescription *attributes, size_t count);
-        assets_id load_pixel_shader(const char *path, const char *name, InputAttributeDescription *attributes, size_t count);
+        assets_id
+        load_vertex_shader(const char *path, const char *name, InputAttributeDescription *attributes, size_t count);
+        assets_id
+        load_pixel_shader(const char *path, const char *name, InputAttributeDescription *attributes, size_t count);
         assets_id load_font(const char *path, const char *name, size_t size);
         int has_assets(const char *name);
         CompositorHandler get_compositor();
@@ -65,8 +67,8 @@ namespace System::Render
                 if (!has_assets(drawIntent.common.pixel_shader))
                 {
                     // TODO: Encode InputAttributeDescription and Count
-                    common.pixel_shader =
-                            load_pixel_shader(drawIntent.common.pixel_shader, drawIntent.common.pixel_shader, nullptr, 0);
+                    common.pixel_shader = load_pixel_shader(
+                            drawIntent.common.pixel_shader, drawIntent.common.pixel_shader, nullptr, 0);
                 }
                 else
                 {
@@ -83,8 +85,8 @@ namespace System::Render
                 if (!has_assets(drawIntent.common.vert_shader))
                 {
                     // TODO: Encode InputAttributeDescription and Count
-                    common.vert_shader =
-                            load_vertex_shader(drawIntent.common.vert_shader, drawIntent.common.vert_shader, nullptr, 0);
+                    common.vert_shader = load_vertex_shader(
+                            drawIntent.common.vert_shader, drawIntent.common.vert_shader, nullptr, 0);
                 }
                 else
                 {
@@ -110,7 +112,7 @@ namespace System::Render
                 special = ComposedSpriteDesc{};
                 auto &sprite = std::get<ComposedSpriteDesc>(special);
                 sprite.texture = get_assets_id(sprite_draw_desc.texture);
-                if (sprite.texture == -1)
+                if (sprite.texture == (uint32_t) -1)
                 {
                     sprite.texture = load_sprite(sprite_draw_desc.texture, sprite_draw_desc.texture, 0, 0);
                 }
@@ -131,7 +133,7 @@ namespace System::Render
                 // text, fontname, fontsize, position, anchor
                 text.text = text_draw_desc.text;
                 text.font = get_assets_id(text_draw_desc.font_name);
-                if (text.font == -1)
+                if (text.font == (uint32_t) -1)
                 {
                     text.font = load_font(text_draw_desc.font_name, text_draw_desc.font_name, 0);
                 }
